% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\name{flowr-package}
\alias{flowr-package}
\alias{flowr}
\title{Transport Modeling}
\description{
\emph{flowr} provides efficient tools for transportation modeling, specifically route
enumeration and traffic assignment tasks. The package implements the path-sized logit model for
traffic assignment and provides utilities for network processing.


\strong{Network Processing}

 \code{\link[=linestrings_to_graph]{linestrings_to_graph()}} --- Convert LINESTRING geometries to graph\cr
 \code{\link[=create_undirected_graph]{create_undirected_graph()}} --- Convert directed graph to undirected\cr
 \code{\link[=consolidate_graph]{consolidate_graph()}} --- Consolidate graph by removing intermediate nodes\cr
 \code{\link[=simplify_network]{simplify_network()}} --- Simplify network by keeping only traversed edges\cr


\strong{Traffic Assignment}

 \code{\link[=run_assignment]{run_assignment()}} --- Run traffic assignment using path-sized logit model\cr


\strong{Graph Utilities}

 \code{\link[=normalize_graph]{normalize_graph()}} --- Normalize node IDs to consecutive integers\cr
 \code{\link[=nodes_from_graph]{nodes_from_graph()}} --- Extract unique nodes from graph\cr
 \code{\link[=linestrings_from_graph]{linestrings_from_graph()}} --- Convert graph to LINESTRING geometries\cr
 \code{\link[=distances_from_graph]{distances_from_graph()}} --- Compute distance matrix from graph\cr


\strong{OD Matrix Utilities}

 \code{\link[=melt_od_matrix]{melt_od_matrix()}} --- Convert origin-destination matrix to long format\cr


\strong{Data}

 \code{\link{network_gcc}} --- Multimodal transport network for the Gulf Cooperation Council (GCC) region\cr
 \code{\link{od_matrices_gcc}} --- Origin-destination matrices for five cargo types in the GCC region\cr
 \code{\link{zones_gcc}} --- Zone locations and descriptions for OD-matrix locations in the GCC region\cr
}
\details{
The package uses efficient C implementations for critical path operations and leverages:
\itemize{
  \item \code{collapse} - Fast data transformations
  \item \code{geodist} - Fast geodesic distance computations
  \item \code{igraph} - Graph operations and shortest path algorithms
}
}
\author{
Sebastian Krantz \email{sebastian.krantz@graduateinstitute.ch} and Kamol Roy \email{kamol.roy08@gmail.com}
}
