% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linestring_to_graph}
\alias{linestring_to_graph}
\title{Convert Linestring to Graph}
\usage{
linestring_to_graph(lines, digits = 6)
}
\arguments{
\item{lines}{An sf data frame of LINESTRING geometries.}
}
\value{
A data.frame representing the graph with columns:
\itemize{
 \item \code{line} - Line identifier
 \item \code{from} - Starting node ID
 \item \code{FX} - Starting node X-coordinate (longitude)
 \item \code{FY} - Starting node Y-coordinate (latitude)
 \item \code{to} - Ending node ID
 \item \code{TX} - Ending node X-coordinate (longitude)
 \item \code{TY} - Ending node Y-coordinate (latitude)
}
}
\description{
Convert Linestring to Graph
}
