% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{print.flowr}
\alias{print.flowr}
\title{Print Flowr Object}
\usage{
\method{print}{flowr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{flowr}, typically returned by \code{\link{run_assignment}}.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print method for objects of class \code{flowr}. Displays a summary of the flow assignment result.
}
